+++ lib/ts/recipe/emailverification/emaildelivery/services/smtp/emailVerify.ts
@@ -41,6 +41,11 @@ export function getEmailVerifyEmailHTML(appName: string, email: string, verifica
             <title>*|MC:SUBJECT|*</title>
 
             <style type="text/css">
+                body {
+                    max-width: 100vw;
+                    overflow: hidden;
+                }
+        
                 p {
                     margin: 10px 0;
                     padding: 0;
@@ -789,6 +794,12 @@ export function getEmailVerifyEmailHTML(appName: string, email: string, verifica
                     }
 
                 }
+                
+                @media only screen and (max-width: 480px) {
+                    #meant-for {
+                        padding: 20px;
+                    }
+                }
             </style>
         </head>
 
@@ -862,7 +873,7 @@ export function getEmailVerifyEmailHTML(appName: string, email: string, verifica
                                                                         <div
                                                                             style="background-color:#fafafa; border-top: 1px solid #ddd; padding-left: 15%; padding-right: 15%; padding-bottom: 24px; padding-top: 24px">
                                                                             <p
-                                                                                style="max-width: 394px !important; font-family: 'Hevetica', sans-serif; font-size: 14px; line-height: 23px; font-weight:400;  text-align: center; color: #808080;">
+                                                                                style="max-width: 600px !important; margin: auto; font-family: 'Hevetica', sans-serif; font-size: 14px; line-height: 23px; font-weight:400;  text-align: center; color: #808080;">
                                                                                 Alternatively, you can directly paste this link
                                                                                 in your browser <br>
                                                                                 <a style="font-family: 'Helvetica', sans-serif, sans-serif; text-align: center; word-break: break-all; font-weight: 400; font-size: 14px; line-height: 23px; color: #007aff !important;"
@@ -908,6 +919,7 @@ export function getEmailVerifyEmailHTML(appName: string, email: string, verifica
 
 
                                                                     <p
+                                                                        id="meant-for"
                                                                         style="font-family: 'Helvetica', sans-serif; font-size: 16px; line-height: 26px; font-weight:400; text-align: center; color: #808080">
                                                                         This email is meant for <a
                                                                             style="font-family: 'Helvetica', sans-serif; text-align: center; word-break: break-all; font-weight: 400; font-size: 16px; line-height: 26px; color: #808080 !important;"